write and read from the disk

+++ Image_Manager/include/io_tools.hpp
@@ -16,8 +16,8 @@ namespace igg::io_tools {
 struct ImageData {
   int rows;
   int cols;
-  uint8_t max_val;
-  std::vector<uint8_t> data;
+  int max_val;
+  std::vector<int> data;
 };
 
 /// Reads from a pgm image from ascii file. Returns empty ImageData if the path
